requirements.txt
pyautogui
pillow
pyscreeze
opencv-python
pytesseract